cloud_trader/training/requirements.txt
datasets>=2.19.1
tensorboard>=2.16.0
google-cloud-bigquery>=3.25.0
